Move bank balance chart from site header to the donation section
Currently the chart showing the bank balance is in the header on every
page, right under the title of the site/project. This frames the whole
project as being about our bank balance—which is in fact a minor part of
our focus.

This commit moves it down into the donations section, where it feels
more thematically at home.

I've also removed all the css for the chart, which is unnecessary in
this new position. The layout and width are maintained by the default of
block elements taking up 100% width of their parent.